+++ core/control/commands/gcode_handler/gcode_handler.c
@@ -310,7 +310,7 @@ int execute_g_command(const unsigned char *command, ssize_t len)
         {
             planner_lock();
             char buf[60];
-            snprintf(buf, 60, "parse error: %.*s", len, command);
+            snprintf(buf, 60, "parse error: %.*s", (int)len, command);
             buf[59] = 0;
             send_error(-1, buf);
             return rc;
